:hammer: Move the cache code into its own module
