+++ src/rogallo/cache.py
@@ -1,0 +1,100 @@
+"""Provides support for a local cache for remote content."""
+
+##############################################################################
+# Python imports.
+from datetime import datetime
+from json import JSONDecodeError, dumps, loads
+from pathlib import Path
+
+##############################################################################
+# BagOfStuff imports.
+from bagofstuff.cache import CacheManager
+
+##############################################################################
+# Wasat imports.
+from wasat import GeminiURI
+
+##############################################################################
+# Local imports.
+from .data.locations import cache_dir
+from .document import Document
+
+
+##############################################################################
+class ContentCache(CacheManager):
+    """A cache manager for remote content."""
+
+    def __init__(self) -> None:
+        """Initialise the content cache."""
+        super().__init__(cache_dir())
+
+    def _cache_files(self, uri: GeminiURI) -> tuple[Path, Path]:
+        """Get the paths to the cache files.
+
+        Args:
+            uri: The URI to get the cache files for.
+
+        Returns:
+            A tuple containing the paths to the cache files.
+        """
+        cache_path = self.get(uri=uri)
+        return cache_path.with_suffix(".meta"), cache_path.with_suffix(".content")
+
+    def get_document(self, uri: GeminiURI) -> Document | None:
+        """Get a cached copy of a document for a given URI.
+
+        Args:
+            uri: The URI to get the cached copy for.
+
+        Returns:
+            The cached document, or `None` if it is not cached.
+        """
+        # TODO: Expiration.
+        meta_data_file, content_file = self._cache_files(uri)
+        try:
+            meta_data = loads(meta_data_file.read_text(encoding="utf-8"))
+        except (OSError, JSONDecodeError):
+            return None
+        try:
+            return Document(
+                location=uri,
+                original_location=GeminiURI(meta_data.get("original_location", uri)),
+                content=content_file.read_text(encoding="utf-8"),
+                mime_type=meta_data.get("mime_type"),
+                from_cache=True,
+            )
+        except OSError:
+            return None
+
+    def add_document(self, document: Document) -> Document:
+        """Add a document to the cache.
+
+        Args:
+            document: The document to cache.
+
+        Returns:
+            The document that was cached.
+        """
+        if not isinstance(document.location, GeminiURI):
+            return document
+        meta_data_file, content_file = self._cache_files(document.location)
+        try:
+            content_file.write_text(document.content, encoding="utf-8")
+            meta_data_file.write_text(
+                dumps(
+                    {
+                        "location": str(document.location),
+                        "original_location": str(document.original_location),
+                        "mime_type": document.mime_type,
+                        "cached_at": datetime.now().isoformat(),
+                    },
+                    indent=4,
+                ),
+                encoding="utf-8",
+            )
+        except OSError:
+            pass
+        return document
+
+
+### cache.py ends here
