+++ src/rogallo/screens/main.py
@@ -3,15 +3,9 @@
 ##############################################################################
 # Python imports.
 from argparse import Namespace
-from datetime import datetime
-from json import JSONDecodeError, dumps, loads
 from mimetypes import guess_type
 from pathlib import Path
 from webbrowser import open as open_in_browser
-
-##############################################################################
-# BagOfStuff imports.
-from bagofstuff.cache import CacheManager
 
 ##############################################################################
 # Pyperclip imports.
@@ -50,6 +44,7 @@ from wasat.uri import GEMINI_PREFIX
 ##############################################################################
 # Local imports.
 from .. import __version__
+from ..cache import ContentCache
 from ..commands import (
     AddLocationToBookmarks,
     Backward,
@@ -89,7 +84,6 @@ from ..data import (
     trust_file,
     update_configuration,
 )
-from ..data.locations import cache_dir
 from ..document import Document
 from ..messages import CopyToClipboard, OpenDocument, OpenLocation, OpenURI
 from ..preflight import (
@@ -232,7 +226,7 @@ class Main(EnhancedScreen[None]):
         super().__init__()
         self._arguments = arguments
         """The command line arguments."""
-        self._cache = CacheManager(cache_dir())
+        self._cache = ContentCache()
         """The disk cache manager."""
 
     def compose(self) -> ComposeResult:
@@ -275,74 +269,6 @@ class Main(EnhancedScreen[None]):
                     self._location_history.current_item.location, from_history=True
                 )
             )
-
-    def _cache_files(self, uri: GeminiURI) -> tuple[Path, Path]:
-        """Get the paths to the cache files.
-
-        Args:
-            uri: The URI to get the cache files for.
-
-        Returns:
-            A tuple containing the paths to the cache files.
-        """
-        cache_path = self._cache.get(uri=uri)
-        return cache_path.with_suffix(".meta"), cache_path.with_suffix(".content")
-
-    def _cached(self, uri: GeminiURI) -> Document | None:
-        """Get a cached copy of a document for a given URI.
-
-        Args:
-            uri: The URI to get the cached copy for.
-
-        Returns:
-            The cached document, or `None` if it is not cached.
-        """
-        # TODO: Expiration.
-        meta_data_file, content_file = self._cache_files(uri)
-        try:
-            meta_data = loads(meta_data_file.read_text(encoding="utf-8"))
-        except (OSError, JSONDecodeError):
-            return None
-        try:
-            return Document(
-                location=uri,
-                original_location=GeminiURI(meta_data.get("original_location", uri)),
-                content=content_file.read_text(encoding="utf-8"),
-                mime_type=meta_data.get("mime_type"),
-                from_cache=True,
-            )
-        except OSError:
-            return None
-
-    def _add_to_cache(self, document: Document) -> Document:
-        """Cache a document.
-
-        Args:
-            document: The document to cache.
-
-        Returns:
-            The document that was cached.
-        """
-        if not isinstance(document.location, GeminiURI):
-            return document
-        meta_data_file, content_file = self._cache_files(document.location)
-        try:
-            content_file.write_text(document.content, encoding="utf-8")
-            meta_data_file.write_text(
-                dumps(
-                    {
-                        "location": str(document.location),
-                        "original_location": str(document.original_location),
-                        "mime_type": document.mime_type,
-                        "cached_at": datetime.now().isoformat(),
-                    },
-                    indent=4,
-                ),
-                encoding="utf-8",
-            )
-        except OSError:
-            pass
-        return document
 
     def check_action(self, action: str, parameters: tuple[object, ...]) -> bool | None:
         """Check if an action is possible to perform right now.
@@ -460,7 +386,7 @@ class Main(EnhancedScreen[None]):
         if self._is_displayable(response.mime_type):
             self.post_message(
                 OpenDocument(
-                    document=self._add_to_cache(
+                    document=self._cache.add_document(
                         Document(
                             location=uri,
                             original_location=request.location,
@@ -519,7 +445,7 @@ class Main(EnhancedScreen[None]):
 
         # If a cached copy of the document exists and the request allows it,
         # use that instead of making a network request.
-        if request.allow_cached and (cached_document := self._cached(uri)):
+        if request.allow_cached and (cached_document := self._cache.get_document(uri)):
             self.post_message(
                 OpenDocument(
                     document=cached_document,
